requirements.txt
streamlit>=1.30.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-whois>=0.9.4
pandas>=2.0.0
